design-ops-navigator/backend/pyproject.toml
[project]
name = "design-ops-navigator-backend"
version = "0.1.0"
description = "Design Ops Navigator — AI UX critique backend (Google ADK)"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "agentops>=0.4.21",
    "beautifulsoup4>=4.14.3",
    "ddgs>=9.11.4",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.135.1",
    "firebase-admin>=7.2.0",
    "google-adk[eval]>=1.27.1",
    "google-cloud-firestore>=2.25.0",
    "google-cloud-storage>=3.9.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "slowapi>=0.1.9",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]
